+++ src/Spreads.Extensions/Storage/Aeron/Logbuffer/TermAppender.cs
@@ -246,7 +246,7 @@ namespace Spreads.Storage.Aeron.Logbuffer {
          * @return the termId in which the append operation took place.
          */
         public static int TermId(long result) {
-            return (int)(result >> 32);
+            return (int)((long)((ulong)result >> 32));
         }
 
         private long HandleEndOfLogCondition(
